% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nl_methods.R
\name{print.summary_nl_fit}
\alias{print.summary_nl_fit}
\title{Print method for summary_nl_fit objects}
\usage{
\method{print}{summary_nl_fit}(x, ...)
}
\arguments{
\item{x}{A \code{summary_nl_fit} object.}

\item{...}{Further arguments passed to \code{print}.}
}
\value{
\code{x} invisibly.
}
\description{
Prints a \code{summary_nl_fit} object returned by
\code{\link{summary.nl_fit}}.
}
